03_real_time_data_communication_part_1/envirement_variables/python/pyproject.toml
[tool.poetry]
name = "python-.env-example"
version = "0.1.0"
description = ""
authors = ["Alex8943 <65241585+Alex8943@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
